#2225 :truck: move obw composable files in ergebnismeldung/OBW
