+++ wls-gui-wahllokalsystem/src/composables/ergebnismeldung/OBW/obwStapelAUtils.ts
