+++ wls-gui-wahllokalsystem/src/composables/ergebnismeldung/OBW/obwStapelBUtils.ts
