+++ wls-gui-wahllokalsystem/src/composables/ergebnismeldung/OBW/obwStapelCUtils.ts
