+++ wls-gui-wahllokalsystem/src/views/auszaehlung/obw/StapelCView.vue
@@ -25,7 +25,7 @@ import { useDisplay } from "vuetify/framework";
 
 import TheOBWStapelCErfassungCard from "@/components/ergebnisermittlung/OBW/stapelC/TheOBWStapelCErfassungCard.vue";
 import TheOBWStapelCSummaryCard from "@/components/ergebnisermittlung/OBW/stapelC/TheOBWStapelCSummaryCard.vue";
-import { useOBWStapelCUtils } from "@/composables/ergebnisermittlung/obwStapelCUtils.ts";
+import { useOBWStapelCUtils } from "@/composables/ergebnismeldung/OBW/obwStapelCUtils.ts";
 import { useUserStore } from "@/stores/userStore.ts";
 
 const { params: routeParams } = useRoute();
